LayoutTests/imported/w3c/web-platform-tests/css/css-typed-om/the-stylepropertymap/properties/flood-opacity-expected.txt


PASS Can set 'flood-opacity' to CSS-wide keywords
FAIL Can set 'flood-opacity' to var() references assert_equals: expected 2 but got 1
FAIL Can set 'flood-opacity' to a number assert_class_string: specified calc must be a CSSMathSum expected "[object CSSMathSum]" but got "[object CSSUnitValue]"
PASS Setting 'flood-opacity' to a length throws TypeError
FAIL Setting 'flood-opacity' to a percent throws TypeError assert_throws_js: function "() => styleMap.set(propertyName, example.input)" did not throw
PASS Setting 'flood-opacity' to a time throws TypeError
PASS Setting 'flood-opacity' to an angle throws TypeError
PASS Setting 'flood-opacity' to a flexible length throws TypeError
PASS Setting 'flood-opacity' to a URL throws TypeError
PASS Setting 'flood-opacity' to a transform throws TypeError
